requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
gunicorn
python-dotenv

